Fix a harmless compiler warning in zipfile.c
FossilOrigin-Name: 60c694c1ab26a7a096f17ccea5a93ecda0f9f2113ab5fdc8b17dbffc787724fc

+++ ext/misc/zipfile.c
@@ -1194,7 +1194,7 @@ static int zipfileGetMode(
 ** nB is the value of strlen(zB). This function returns 0 if the strings are
 ** identical, ignoring any trailing '/' character in either path.  */
 static int zipfileComparePath(const char *zA, const char *zB, int nB){
-  int nA = strlen(zA);
+  int nA = (int)strlen(zA);
   if( zA[nA-1]=='/' ) nA--;
   if( zB[nB-1]=='/' ) nB--;
   if( nA==nB && memcmp(zA, zB, nA)==0 ) return 0;
@@ -1525,4 +1525,3 @@ int sqlite3_zipfile_init(
   (void)pzErrMsg;  /* Unused parameter */
   return zipfileRegister(db);
 }
-
